Update TokenGateway.php
This allows the setDemoMode() feature with TwoCheckoutPlus_Token. It's pretty important. By default, the sandbox has the mode set to "Parameter" and looks for the JSON to include a "demo" property to be passed in so that transactions will not say "Authorization Failed" in the sandbox. Therefore, when testing in the sandbox, one needs to do setTestMode(true) as well as setDemoMode(true). With setTestMode(true), you at least get to use the sandbox API. But with setDemoMode(true), you can use the test cards and they will validate without saying "Authorization Failed" on them.